implemented given changes

+++ CSV/src/main/java/org/example/csv/csv/controller/CSVToEdiController.java
@@ -24,18 +24,21 @@ public class CSVToEdiController {
     private CSVToEdiServices csvToEdiServices;
 
     @GetMapping(value = "/csv-to-edi846", produces = "text/edi")
-    public ResponseEntity<Object> csvToEdi(@RequestParam(value = "csvFile") MultipartFile csvFile, HttpServletResponse response) {
-        File responseFile;
+    public ResponseEntity<Object> csvToEdi(@RequestParam(value = "csvFile") MultipartFile csvFile, HttpServletResponse response) throws IOException {
+        File responseFile = null;
         try {
             responseFile = csvToEdiServices.csvToEdiConverter(csvFile);
             byte[] responseFileByte = Files.readAllBytes(responseFile.toPath());
             response.setHeader(HttpHeaders.CONTENT_DISPOSITION,
                     "attachment; filename=\"" + responseFile.getName() + "\"");
-            FileUtils.delete(responseFile);
             FileUtils.delete(new File(csvFile.getOriginalFilename()));
             return ResponseEntity.status(HttpStatus.OK).body(responseFileByte);
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+        } finally {
+            if (responseFile != null) {
+                FileUtils.delete(responseFile);
+            }
         }
     }
 }
